requirements.txt
black==22.8.0
Django~=3.2.7
django-coverage-plugin==2.0.3
django-dynamic-fixture==3.1.2
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-filters==0.10.2
django-test-without-migrations==0.6
flake8==4.0.1
gevent
mkdocs==1.1.2
mkdocs-material==7.0.3
pre-commit==2.12.1
pylint==2.8.2
pylint-django==2.4.4
pytest==7.1.3
pytest-django==4.5.2
pytest-xdist==2.5.0
pytest-cov==4.0.0
python-decouple==3.4
psycopg2-binary==2.9.3  # PostgreSQL adapter
whitenoise==6.2.0       # For static files
gunicorn==20.1.0        # Production WSGI server
requests==2.28.1        # For Rancher metadata checks
django-health-check==3.16.5  # Optional: comprehensive health checks
